requirements.txt
insightface
opencv-python
numpy
scikit-learn
matplotlib
tensorflow
onnxruntime
ultralytics
mysql-connector-python
streamlit